move the currency formatter to the lib folder
