+++ src/lib/format-currency.js
